Lint:fix lints

+++ pkg/cli/util/util.go
@@ -170,5 +170,5 @@ func HumanDuration(d time.Duration) string {
 	} else if hours < 24*365*8 {
 		return fmt.Sprintf("%dy%dd", hours/24/365, (hours/24)%365)
 	}
-	return fmt.Sprintf("%dy", int(hours/24/365))
+	return fmt.Sprintf("%dy", hours/24/365)
 }
